ViewとStateの整理
